rename query-string-service
